Restructure bulk benchmarks
In order to extend the bulk benchmark further, this change moves most
common functions out of the bulk.rs binary and into a lib.rs and stats.rs file.

Functionally there are no relevant changes.
