+++ bench/Cargo.toml
@@ -18,7 +18,3 @@ structopt = "0.3"
 tokio = { version = "1.0.1", features = ["rt"] }
 tracing = "0.1.10"
 tracing-subscriber = { version = "0.2.5", default-features = false, features = ["env-filter", "fmt", "ansi", "chrono"]}
-
-[[bin]]
-name = "bulk"
-path = "src/bulk.rs"
